Move to upload types to appropriate file, remove duplicate JSONL upload component.
